requirements.txt
tensorflow==2.21.0
scikit-learn==1.8.0
pandas==2.3.3
numpy==2.4.3
matplotlib==3.10.8
seaborn==0.13.2
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings==2.9.1
streamlit==1.55.0
joblib==1.5.3
xgboost==3.2.0
shap==0.51.0
scipy==1.17.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
prometheus-client==0.22.1
gunicorn==23.0.0
python-multipart==0.0.22
